+++ docs/source/history.md
@@ -544,5 +544,9 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 ## v0.7.4
 
+- Input:
+  - Adding output options to `windio`
+- Output:
+  - New output `StateTurbineTable`, exporting state-turbine data to NetCDF
   
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v0.7.4](https://github.com/FraunhoferIWES/foxes/commits/v0.7.4)